roadmap/requirements.txt

streamlit
python-dotenv 
google.generativeai 
streamlit_mermaid 